Test data from Florida Bay

Small data set collected for positioning using acoustic telemetry and YAPS.
The data are part of a feasibility study using YAPS on Vemco PPM style data to track fish in shallow parts of Florida Bay. Data were collected using VR2 (Vemco) hydrophones.
Included in yaps with permission from J.S. Rehage, FIU Florida International University.

ssu1

Format

A list containing 3 data.tables:

hydros

serial Hydrophone serial number.

x,y,z Position of hydrophones in UTM.

sync_tag ID of co-located sync tag. Must be identical to entries in data.table detections$tag.

idx Unique values from 1:nrow(hydros).

detections

ts Timestamp of detection in POSIXct().

tag ID of detected tag.

epo Timestamp as number of seconds since Unix epoch. Can be obtained using as.numeric(ts).

frac Sub-second part of detection timestamp in fractions of second [0-1].

serial Serial number of detecting hydrophone. Must match entry in data.table hydros.

gps

ts Timestamp of gps position in POSIXct().

utm_x, utm_y Coordinates of position. Same projection and coordinate system as used in hydros.